+++ Tests/DependencyInjection/IndragunawanApiRateLimitExtensionTest.php
@@ -30,7 +30,7 @@ class IndragunawanApiRateLimitExtensionTest extends TestCase
     /**
      * {@inheritdoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->container = new ContainerBuilder();
         $this->container->setParameter('kernel.cache_dir', '../../var/cache');
@@ -41,7 +41,7 @@ class IndragunawanApiRateLimitExtensionTest extends TestCase
     /**
      * {@inheritdoc}
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         unset($this->container, $this->extension);
     }
@@ -77,22 +77,6 @@ class IndragunawanApiRateLimitExtensionTest extends TestCase
 
         $storageDefinition = $this->container->getDefinition('indragunawan_api_rate_limit.service.rate_limit_handler');
         $this->assertSame('custom_cache', (string) $storageDefinition->getArgument(0));
-    }
-
-    public function testDeprecatedServiceConfig()
-    {
-        $config = [
-            [
-                'storage' => 'custom_storage',
-            ],
-        ];
-
-        $this->extension->load($config, $this->container);
-
-        $this->assertTrue($this->container->hasDefinition('indragunawan_api_rate_limit.service.rate_limit_handler'));
-
-        $storageDefinition = $this->container->getDefinition('indragunawan_api_rate_limit.service.rate_limit_handler');
-        $this->assertSame('custom_storage', (string) $storageDefinition->getArgument(0)->getArgument(0));
     }
 
     public function testSortFirstMatch()