+++ Tests/EventListener/HeaderModificationListenerTest.php
@@ -13,20 +13,36 @@ namespace Indragunawan\ApiRateLimitBundle\Tests\EventListener;
 
 use Indragunawan\ApiRateLimitBundle\EventListener\HeaderModificationListener;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
+use Symfony\Component\HttpFoundation\ResponseHeaderBag;
+use Symfony\Component\HttpKernel\Event\ResponseEvent;
+use Symfony\Component\HttpKernel\HttpKernelInterface;
 
 class HeaderModificationListenerTest extends TestCase
 {
+    private $kernel;
+
+    public function setUp(): void
+    {
+        $this->kernel = $this->createMock(HttpKernelInterface::class);
+    }
+
+    public function tearDown(): void
+    {
+        $this->kernel = null;
+    }
+
     public function testNotDisplayHeader()
     {
-        $event = $this->getMockBuilder(FilterResponseEvent::class)
-            ->disableOriginalConstructor()
-            ->getMock();
+        $responseHeaderBag = $this->prophesize(ResponseHeaderBag::class);
+        $responseHeaderBag->set()->shouldNotHaveBeenCalled();
 
-        $event->expects($this->never())
-            ->method('getRequest');
+        $response = $this->prophesize(Response::class);
+        $response->headers = $responseHeaderBag->reveal();
+
+        $event = new ResponseEvent($this->kernel, new Request(), HttpKernelInterface::MASTER_REQUEST, $response->reveal());
 
         $headerConfig = [
             'display' => false,
@@ -38,21 +54,13 @@ class HeaderModificationListenerTest extends TestCase
 
     public function testNoRateLimitInfoAttributes()
     {
-        $event = $this->getMockBuilder(FilterResponseEvent::class)
-            ->disableOriginalConstructor()
-            ->getMock();
+        $attributes = $this->prophesize(ParameterBag::class);
+        $attributes->get('_api_rate_limit_info', null)->willReturn(null)->shouldBeCalledTimes(1);
 
-        $resetTime = gmdate('U');
+        $request = $this->prophesize(Request::class);
+        $request->attributes = $attributes->reveal();
 
-        $request = Request::create('/api/me');
-        $request->attributes->set('_api_rate_limit_info', null);
-
-        $event->expects($this->once())
-            ->method('getRequest')
-            ->will($this->returnValue($request));
-
-        $event->expects($this->never())
-            ->method('getResponse');
+        $event = new ResponseEvent($this->kernel, $request->reveal(), HttpKernelInterface::MASTER_REQUEST, new Response());
 
         $headerConfig = [
             'display' => true,
@@ -64,27 +72,21 @@ class HeaderModificationListenerTest extends TestCase
 
     public function testSetResponseHeaders()
     {
-        $event = $this->getMockBuilder(FilterResponseEvent::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
         $resetTime = gmdate('U');
 
-        $request = Request::create('/api/me');
-        $request->attributes->set('_api_rate_limit_info', [
+        $attributes = $this->prophesize(ParameterBag::class);
+        $attributes->get('_api_rate_limit_info', null)->willReturn([
             'limit' => 60,
             'remaining' => 59,
             'reset' => $resetTime,
-        ]);
+        ])->shouldBeCalledTimes(1);
 
-        $event->expects($this->once())
-            ->method('getRequest')
-            ->will($this->returnValue($request));
+        $request = $this->prophesize(Request::class);
+        $request->attributes = $attributes->reveal();
 
         $response = new Response();
-        $event->expects($this->once())
-            ->method('getResponse')
-            ->will($this->returnValue($response));
+
+        $event = new ResponseEvent($this->kernel, $request->reveal(), HttpKernelInterface::MASTER_REQUEST, $response);
 
         $headerConfig = [
             'display' => true,